requirements.txt
# Core Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Image Processing
opencv-python>=4.5.0
scikit-image>=0.18.0
Pillow>=8.3.0
tifffile>=2021.7.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0
keras>=2.6.0

# Graph Neural Networks
torch-geometric>=2.0.0
networkx>=2.6.0

# Segmentation
scikit-learn>=0.24.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Web Framework
flask>=2.0.0
streamlit>=1.0.0

# Utilities
tqdm>=4.62.0
PyYAML>=5.4.0
jupyter>=1.0.0
